updated definition of rhs
Updated plotting conditions and adjusted parameters for major stride and compression. Added checks for plotting and modified return values in the rhs function. Determined the coarseness of the piecewise linear A_of_t() (major_stride size) solely determines the accuracy of the integrator. Next step is making A_of_t() a method of a class and using splines (maybe cubic) to smoothly interpolate with the intent of increasing major_stride beyond the trivial (3 or 4).